Changed where the grid is displayed.
Currently it would conflict with any existing styles applied to the
after pseudo element. Lets say you have a container which has two
columns, neat will also apply a clearfix on it. When a grid is added,
the grid will never show because of the display: table being applied by
the clearfix which effectively hides the grid. Whats more is that since
the grid was positioned absolute, it will also rely on the container to
have a position relative on it - which can cause issues when debugging.
My proposal applies the grid as a background-image and doesn’t require
to element to have position relative nor does it need to be positioned
absolute.